pypi pkg/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mangadex-sync"
version = "2.0.3"
description = "Export and import your MangaDex library to/from MyAnimeList and AniList — with scores, statuses, and chapter progress"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Spider2742", email = "mail@spidey.in" }
]
keywords = ["mangadex", "myanimelist", "anilist", "manga", "export", "import", "sync", "mal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
    "Topic :: Utilities",
]
dependencies = [
    "pywebview>=5.0.0",
    "PyQt6>=6.4.0",
    "PyQt6-WebEngine>=6.4.0",
    "qtpy>=2.4.0",
    "flask>=3.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = ["build", "twine"]

[project.scripts]
mangadex-sync = "mangadex_sync:main"

[project.urls]
Homepage = "https://github.com/spider2742/mangadex-sync"
"Bug Tracker" = "https://github.com/spider2742/mangadex-sync/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mangadex_sync*"]